refact: solving conflicts
