+++ src/app.js
@@ -1,7 +1,6 @@
 const express = require('express');
 require('express-async-errors');
 const passport = require('passport');
-const GoogleStrategy = require('passport-google-oauth20').Strategy;
 const cookieParser = require('cookie-parser');
 const { encryptCookieNodeMiddleware } = require('encrypt-cookie');
 const { expressjwt: jwt } = require('express-jwt');
@@ -12,7 +11,6 @@ const logger = require('./utils/logger');
 const errorHandler = require('./middlewares/error');
 require('dotenv').config();
 const { connectToMongoAtlas } = require('./db/connection');
-const User = require('./models/user');
 
 const app = express();
 app.use(express.json());
@@ -30,9 +28,6 @@ require('./middlewares/passport-auth');
 app.use(passport.initialize());
 
 app.use('/api', routes);
-// app.use('/', (req, res, next) => {
-//   res.send('Hello World!');
-// });
 
 app.use(errorHandler);
